Display the existing family members using GET in ExistingUser.jsx
